pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-prishtina-milvus-client"
version = "0.1.0"
authors = [
    { name = "AI Prishtina", email = "info@albanmaxhuni.com" },
]
description = "A Python client for Milvus vector database with advanced features for AI applications"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
dependencies = [
    "pymilvus>=2.3.0",
    "pyyaml>=6.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pydantic>=1.10.0,<2.0.0",
    "boto3>=1.34.0",
    "google-cloud-storage>=2.0.0",
    "azure-storage-blob>=12.0.0",
    "azure-identity>=1.0.0",
    "requests>=2.0.0",
    "google-cloud-aiplatform>=1.36.0",
    "confluent-kafka>=2.3.0",
    "pytesseract>=0.3.10",
    "pdf2image>=1.16.3",
    "Pillow>=10.0.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "sentence-transformers>=2.2.2",
    "redis>=5.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0"
]

[project.urls]
Homepage = "https://github.com/yourusername/ai-prishtina-milvus-client"
Documentation = "https://github.com/yourusername/ai-prishtina-milvus-client/docs"
Repository = "https://github.com/yourusername/ai-prishtina-milvus-client.git"
Issues = "https://github.com/yourusername/ai-prishtina-milvus-client/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=ai_prishtina_milvus_client"

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 100
target-version = "py38"
select = ["E", "F", "B", "I"]
ignore = []

[tool.coverage.run]
source = ["ai_prishtina_milvus_client"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
] 